backend/recommend-server/requirements.txt
beautifulsoup4==4.12.3
konlpy==0.6.0
matplotlib==3.8.4
matplotlib-inline==0.1.6
numpy==1.26.4
openai==1.44.0
pandas==2.2.2
requests==2.32.2
requests-file==1.5.1
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.1
seaborn==0.13.2
selenium==4.24.0
wordcloud==1.9.3
sqlalchemy
pyyaml
pymysql
pymongo
fastapi
uvicorn
